requirement.txt
boto3>=1.5
Flask
flask-sqlalchemy
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests